refactor(cli): move generate command into dedicated file
